functions/ThresholdEvaluator/requirements.txt
azure-functions
azure-eventgrid
pyodbc
pandas
requests
python-dotenv
openai
